improve code base and tests
