+++ cms_form/tests/__init__.py
@@ -1,1 +1,3 @@
-from . import test_form
+from . import test_form_base
+from . import test_form_cms
+from . import test_form_render
